pyproject.toml
[project]
name = "homepage"
version = "0.10.0"
description = "My homepage"
readme = "README.md"
requires-python = ">=3.13"
dependencies = []

[dependency-groups]
dev = ["numpy>=2.4.3", "pillow>=12.1.1", "playwright>=1.58.0", "ruff>=0.15.6"]
